requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
asgiref
attrs
Django
django-bootstrap5
django-cleanup
django-htmx
django-ranged-response
django-sass-processor
django-simple-captcha
django-utils-six
django-widget-tweaks
frozenlist
idna
libsass
Markdown
multidict
pillow
propcache
python-dotenv
python-slugify
pytils
sqlparse
telepot
text-unidecode
typus
tzdata
urllib3
yarl
